refactor: remove all Apollo Client infrastructure
Remove the entire Apollo Client dependency tree from the frontend:

Core infrastructure deleted from @webiny/app:
- ApolloLinkPlugin, ApolloDynamicLink, ApolloCacheObjectIdPlugin
- AddQuerySelectionPlugin, ConsoleLinkPlugin, OmitTypenameLinkPlugin
- TenantHeaderLinkPlugin, NetworkErrorLinkPlugin
- apollo-client/ directory (InMemoryCache, IntrospectionFragmentMatcher)
- All Apollo npm packages (apollo-client, apollo-link-*, apollo-cache-*)

Bridge removed from @webiny/app-admin:
- ApolloProvider wrapper (createApolloProvider)
- AddGraphQLQuerySelection component
- createApolloClient prop from AdminProps
- TenantHeaderLinkPlugin registration from TenancyProvider
  (already handled by DI-based GraphQLClientDecorator)

Wiring removed from @webiny/app-serverless-cms:
- apolloClientFactory.ts and apolloLinks.ts deleted
- Container/DI import removed from Admin

Leaf consumers removed:
- app-headless-cms: getObjectId plugin and getFetchPolicy (both dead)
- api-core: x-apollo-tracing CORS headers

Error overlay UI components moved from plugins/NetworkErrorLinkPlugin/
to errors/overlay/ (still used by ErrorOverlayNetworkErrorHandler).
