+++ packages/app-admin/package.json
@@ -14,7 +14,6 @@
   "author": "Pavel Denisjuk",
   "license": "MIT",
   "dependencies": {
-    "@apollo/react-components": "^3.1.5",
     "@emotion/css": "^11.13.5",
     "@emotion/react": "11.14.0",
     "@emotion/styled": "11.14.1",
@@ -42,11 +41,6 @@
     "@webiny/utils": "0.0.0",
     "@webiny/validation": "0.0.0",
     "@webiny/wcp": "0.0.0",
-    "apollo-cache": "^1.3.5",
-    "apollo-client": "^2.6.10",
-    "apollo-link": "^1.2.14",
-    "apollo-link-context": "^1.0.20",
-    "apollo-utilities": "^1.3.4",
     "bytes": "^3.1.2",
     "classnames": "^2.5.1",
     "graphql": "^16.14.2",
