+++ packages/app-admin/src/base/Admin.tsx
@@ -1,9 +1,6 @@
 import React from "react";
-import type { Container } from "@webiny/di";
 import { plugins } from "@webiny/plugins";
 import { App, DiContainerProvider } from "@webiny/app";
-import type { ApolloClientFactory } from "./providers/ApolloProvider.js";
-import { createApolloProvider } from "./providers/ApolloProvider.js";
 import { Base } from "./Base.js";
 import { createUiStateProvider } from "./providers/UiStateProvider.js";
 import { createAdminUiStateProvider } from "./providers/AdminUiStateProvider.js";
@@ -14,7 +11,6 @@ import { createRootContainer } from "~/base/createRootContainer.js";
 import { WcpProvider } from "~/presentation/wcp/WcpProvider.js";
 import { createTenancyProvider } from "~/presentation/tenancy/createTenancyProvider.js";
 import { TelemetryAdminAppStart } from "./TelemetryAdminAppStart.js";
-import { ApolloClientFeature } from "~/features/apolloClient/feature.js";
 import { SecurityFeature } from "~/features/security/SecurityFeature.js";
 import { FormModelFeature } from "~/features/formModel/feature.js";
 import type { PluginCollection } from "@webiny/plugins/types.js";
@@ -24,26 +20,20 @@ import { ListPresenterFeature } from "~/presentation/listPresenter/index.js";
 import { NotificationsRenderer } from "~/features/notifications/NotificationsRenderer.js";
 
 export interface AdminProps {
-    createApolloClient: ApolloClientFactory;
-    createLegacyPlugins: (container: Container) => PluginCollection;
+    createLegacyPlugins: () => PluginCollection;
     children?: React.ReactNode;
 }
 
 const container = createRootContainer();
 
-export const Admin = ({ children, createApolloClient, createLegacyPlugins }: AdminProps) => {
-    const uri = process.env.REACT_APP_GRAPHQL_API_URL as string;
-    const apolloClient = createApolloClient({ uri });
+export const Admin = ({ children, createLegacyPlugins }: AdminProps) => {
+    plugins.register(...createLegacyPlugins());
 
-    plugins.register(...createLegacyPlugins(container));
-
-    ApolloClientFeature.register(container, apolloClient);
     SecurityFeature.register(container);
     FormModelFeature.register(container);
     WebinySdkFeature.register(container);
     ListPresenterFeature.register(container);
 
-    const ApolloProvider = createApolloProvider(apolloClient);
     const UIProviders = createUiProviders();
     const UiStateProvider = createUiStateProvider();
     const AdminUiStateProvider = createAdminUiStateProvider();
@@ -53,28 +43,26 @@ export const Admin = ({ children, createApolloClient, createLegacyPlugins }: Adm
     return (
         <DiContainerProvider container={container}>
             <TelemetryAdminAppStart />
-            <ApolloProvider>
-                <WcpProvider>
-                    <App
-                        plugins={[AdminConfigPlugin]}
-                        routes={[]}
-                        providers={[
-                            AdminConfigProvider,
-                            UIProviders,
-                            UiStateProvider,
-                            DialogsProvider,
-                            IconPickerConfigProvider,
-                            AdminUiStateProvider,
-                            TenancyProvider
-                        ]}
-                    >
-                        <Base />
-                        <DefaultIcons />
-                        <NotificationsRenderer />
-                        {children}
-                    </App>
-                </WcpProvider>
-            </ApolloProvider>
+            <WcpProvider>
+                <App
+                    plugins={[AdminConfigPlugin]}
+                    routes={[]}
+                    providers={[
+                        AdminConfigProvider,
+                        UIProviders,
+                        UiStateProvider,
+                        DialogsProvider,
+                        IconPickerConfigProvider,
+                        AdminUiStateProvider,
+                        TenancyProvider
+                    ]}
+                >
+                    <Base />
+                    <DefaultIcons />
+                    <NotificationsRenderer />
+                    {children}
+                </App>
+            </WcpProvider>
         </DiContainerProvider>
     );
 };
