--- packages/app-admin/src/features/apolloClient/abstraction.ts
@@ -1,8 +1,0 @@
-import type { ApolloClient as IApolloClient } from "apollo-client";
-import { createAbstraction } from "@webiny/feature/admin";
-
-export const ApolloClient = createAbstraction<IApolloClient<any>>("ApolloClient");
-
-export namespace ApolloClient {
-    export type Interface = IApolloClient<any>;
-}
