+++ packages/app-headless-cms/package.json
@@ -46,10 +46,6 @@
     "@webiny/react-properties": "0.0.0",
     "@webiny/utils": "0.0.0",
     "@webiny/validation": "0.0.0",
-    "apollo-cache": "^1.3.5",
-    "apollo-client": "^2.6.10",
-    "apollo-link": "^1.2.14",
-    "apollo-utilities": "^1.3.4",
     "dot-prop-immutable": "^2.1.1",
     "graphql": "^16.14.2",
     "lodash": "^4.18.1",
