--- packages/app-headless-cms/src/admin/plugins/getObjectId.ts
@@ -1,10 +1,0 @@
-import { ApolloCacheObjectIdPlugin } from "@webiny/app";
-
-export default new ApolloCacheObjectIdPlugin(obj => {
-    switch (obj.__typename) {
-        case "CmsContentModel":
-            return `${obj.__typename}:${obj.modelId}`;
-        default:
-            return undefined;
-    }
-});
