+++ packages/app-headless-cms/src/admin/plugins/index.ts
@@ -1,6 +1,5 @@
 import icons from "./icons.js";
 import defaultBar from "./editor/defaultBar/index.js";
 import formSettings from "./editor/formSettings/index.js";
-import getObjectId from "./getObjectId.js";
 
-export default () => [icons, defaultBar, formSettings, getObjectId];
+export default () => [icons, defaultBar, formSettings];
