+++ packages/app-serverless-cms/package.json
@@ -35,16 +35,10 @@
     "@webiny/app-websockets": "0.0.0",
     "@webiny/app-workflows": "0.0.0",
     "@webiny/background-tasks": "0.0.0",
-    "@webiny/di": "^1.0.2",
     "@webiny/lexical-editor-actions": "0.0.0",
     "@webiny/lexical-theme": "0.0.0",
     "@webiny/plugins": "0.0.0",
     "@webiny/webhooks": "0.0.0",
-    "apollo-cache": "^1.3.5",
-    "apollo-client": "^2.6.10",
-    "apollo-link": "^1.2.14",
-    "apollo-link-batch-http": "^1.2.14",
-    "apollo-utilities": "^1.3.4",
     "react": "18.3.1",
     "react-dom": "18.3.1"
   },
