+++ packages/app/package.json
@@ -27,24 +27,13 @@
     "@webiny/plugins": "0.0.0",
     "@webiny/react-composition": "0.0.0",
     "@webiny/react-properties": "0.0.0",
-    "@webiny/stdlib": "^0.0.8",
-    "apollo-cache": "^1.3.5",
-    "apollo-cache-inmemory": "^1.6.6",
-    "apollo-client": "^2.6.10",
-    "apollo-link": "^1.2.14",
-    "apollo-link-context": "^1.0.20",
-    "apollo-link-error": "^1.1.13",
-    "apollo-link-http-common": "^0.2.16",
-    "apollo-utilities": "^1.3.4",
     "graphql": "^16.14.2",
     "history": "^5.3.0",
     "invariant": "^2.2.4",
     "lodash": "^4.18.1",
     "mobx": "^6.16.1",
-    "nanoid": "^5.1.15",
     "react": "18.3.1",
     "react-dom": "18.3.1",
-    "ts-invariant": "^0.10.3",
     "warning": "^4.0.3",
     "zod": "4.4.3"
   },
